q2_automation_poc/requirements.txt
playwright
python-dotenv
pandas

